deleted some obsolete directories, imag_segmentation in class
