+++ PIPELINE/applyMask_pipeline.py
